correccion en auth y users

+++ routes/users.js
@@ -3,7 +3,7 @@ const db = require("../models/index");
 const User = db.sequelize.models.User;
 const IsAuthenticated = require('../helpers/auth/isAuthenticated');
 const generateToken = require("../helpers/auth/generateToken");
-const { IsAdmin } = require('../helpers/auth/isAdmin')
+const IsAdmin = require('../helpers/auth/isAdmin')
 
 
 const router = express.Router();
@@ -12,7 +12,7 @@ const router = express.Router();
 router.get("/", IsAuthenticated, IsAdmin, async (req, res) => {
   try {
     const dataUsers = await User.findAll();
-    res.status(200).json({ 
+    res.status(200).json({
       message: "Estos son los usuarios de la base de datos",
       data: dataUsers
     })
@@ -29,7 +29,7 @@ router.get("/:id", async (req, res) => {
 
     const response = { firstName, lastName, email, roleId };
 
-    res.json({data: response});
+    res.json({ data: response });
   } catch (e) {
     res.status(404).json({ message: e.message });
   }
@@ -39,8 +39,8 @@ router.get("/:id", async (req, res) => {
 
 router.put("/:id", IsAuthenticated, async (req, res) => {
   try {
-    if(req.user.userId == req.params.id || req.user.roleId === 1) {
-      if(req.user.roleId === 1){
+    if (req.user.userId == req.params.id || req.user.roleId === 1) {
+      if (req.user.roleId === 1) {
         const response = await User.update(
           {
             firstName: req.body.firstName,
@@ -49,14 +49,14 @@ router.put("/:id", IsAuthenticated, async (req, res) => {
             roleId: req.body.roleId
           },
           {
-            where: { userId: req.params.id}, 
+            where: { userId: req.params.id },
           });
-          if(response){
-            const token = generateToken(req.body)
-            res.json({
-              message: "Profile updated",
-              token: token
-            })
+        if (response) {
+          const token = generateToken(req.body)
+          res.json({
+            message: "Profile updated",
+            token: token
+          })
         } else {
           res.json({ response })
         }
@@ -68,15 +68,15 @@ router.put("/:id", IsAuthenticated, async (req, res) => {
             email: req.body.email
           },
           {
-            where: { userId: req.params.id }, 
+            where: { userId: req.params.id },
           }
         )
-        if(response){
-            const token = generateToken(req.body)
-            res.json({
-              message: "Profile updated",
-              token: token
-            })
+        if (response) {
+          const token = generateToken(req.body)
+          res.json({
+            message: "Profile updated",
+            token: token
+          })
         } else {
           res.json({ response })
         }
@@ -85,7 +85,7 @@ router.put("/:id", IsAuthenticated, async (req, res) => {
       res.status(403).json({ message: "You are not allowed to be here!" })
     }
   } catch (e) {
-    res.status(400).json({ message: e.message});
+    res.status(400).json({ message: e.message });
   }
 });
 
@@ -94,16 +94,18 @@ router.put("/:id", IsAuthenticated, async (req, res) => {
 
 router.delete("/:id", IsAuthenticated, async (req, res) => {
   try {
-    let query = await User.findByPk(req.params.id);
-
-    if (!query) {
-      //Si no encuentra el usuario no existe
-      res.status(400).json({ message: "User not found" })
+    if (req.params.id === req.user.userId || req.user.roleId === 1) {
+      let query = await User.findByPk(req.params.id);
+      if (!query) {
+        //Si no encuentra el usuario no existe
+        res.status(400).json({ message: "User not found" })
+      } else {
+        const response = await User.destroy({
+          where: { userId: req.params.id },
+        });
+        res.json(response);
+      }
     }
-    const response = await User.destroy({
-      where: { userId: req.params.id },
-    });
-    res.json(response);
   } catch (e) {
     res.status(404).json({ message: e.message });
   }
